requirements.txt
flask~=2.2.3
flask-restful
pandas
opencv-python~=4.7.0.72
numpy==1.24.2
Werkzeug~=2.2.3
imageio~=2.27.0
Pillow==9.5.0
gevent~=22.10.2
keras~=2.12.0
sklearn~=0.0.post1
scikit-learn~=1.2.2
tensorflow~=2.12.0
os